+++ gitsvg/cli/_cli.py
@@ -4,6 +4,7 @@ import click
 
 from gitsvg import __version__
 from gitsvg.cli._errors import errors_command
+from gitsvg.cli._layout import layout_command
 from gitsvg.cli._render import render_command
 from gitsvg.cli._schema import schema_command
 from gitsvg.cli._state import state_command
@@ -21,3 +22,4 @@ cli.add_command(errors_command)
 cli.add_command(validate_command)
 cli.add_command(render_command)
 cli.add_command(state_command)
+cli.add_command(layout_command)